src/celcat/scraper.py
"""
CELCAT timetable scraper module.
"""

from bs4 import BeautifulSoup
import logging
from datetime import datetime, timedelta
from typing import List, Dict, Any

from .auth import CelcatAuth
from ..config import STUDENT_ID, CELCAT_BASE_URL

logger = logging.getLogger(__name__)

class CelcatScraper:
    """Handle scraping of timetable data from CELCAT."""
    
    def __init__(self, auth: CelcatAuth):
        """Initialize the CELCAT scraper.
        
        Args:
            auth (CelcatAuth): Authenticated CELCAT session.
        """
        self.auth = auth
        
    def get_timetable_url(self, start_date: datetime) -> str:
        """Generate the timetable URL for a given date.
        
        Args:
            start_date (datetime): The start date for the timetable view.
            
        Returns:
            str: The timetable URL.
        """
        date_str = start_date.strftime("%Y-%m-%d")
        return f"{CELCAT_BASE_URL}/cal?vt=agendaWeek&dt={date_str}&et=student&fid0={STUDENT_ID}"
        
    def extract_events(self, html_content: str) -> List[Dict[str, Any]]:
        """Extract event data from the timetable HTML.
        
        Args:
            html_content (str): The HTML content of the timetable page.
            
        Returns:
            List[Dict[str, Any]]: List of extracted events with their details.
        """
        soup = BeautifulSoup(html_content, 'html.parser')
        events = []
        
        # TODO: Implement event extraction logic based on actual HTML structure
        # This is a placeholder that will need to be updated after inspecting the actual page
        
        try:
            # Example structure (to be updated):
            event_elements = soup.find_all('div', class_='event')
            
            for element in event_elements:
                event = {
                    'title': element.get('title', ''),
                    'start_time': element.get('data-start', ''),
                    'end_time': element.get('data-end', ''),
                    'location': element.get('data-location', ''),
                    'description': element.get('data-description', '')
                }
                events.append(event)
                
        except Exception as e:
            logger.error(f"Error extracting events: {str(e)}")
            
        return events
        
    def get_timetable(self, start_date: datetime = None) -> List[Dict[str, Any]]:
        """Get the timetable for a given date range.
        
        Args:
            start_date (datetime, optional): The start date. Defaults to current date.
            
        Returns:
            List[Dict[str, Any]]: List of timetable events.
        """
        if not start_date:
            start_date = datetime.now()
            
        try:
            url = self.get_timetable_url(start_date)
            self.auth.driver.get(url)
            
            # Wait for the timetable to load
            # TODO: Add appropriate wait condition based on actual page structure
            
            html_content = self.auth.driver.page_source
            return self.extract_events(html_content)
            
        except Exception as e:
            logger.error(f"Error getting timetable: {str(e)}")
            return [] 